--- src/components/layouts/PaginationSection/index.ts
@@ -1,1 +1,0 @@
-export { default as PaginationSection } from './PaginationSection';
